将大部分 int 换成 uint

+++ esurfing/inc/utils/Shutdown.h
@@ -1,11 +1,13 @@
 #ifndef ESURFINGCLIENT_SHUTDOWN_H
 #define ESURFINGCLIENT_SHUTDOWN_H
 
+#include <stdint.h>
+
 /**
  * 关闭函数
  * @param exitCode 退出码
  */
-void shut(int exitCode);
+void shut(uint8_t exitCode);
 
 /**
  * 初始化关闭函数
